requirements.txt
Flask
Pillow
torch
transformers
